packages/core_utils/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "core_utils"
version = "0.1.0"
dependencies = [
    "redis>=5.0.0",
    "python-jose[cryptography]" # For JWT handling
]
